feat: allow slice9'ed drawSprite
