+++ src/ecs/components/draw/sprite.ts
@@ -6,17 +6,14 @@ import {
     type SpriteAnim,
     type SpriteData,
 } from "../../../assets/sprite";
-import { DEF_ANCHOR } from "../../../constants/general";
 import { KEvent, type KEventController } from "../../../events/events";
 import { getRenderProps } from "../../../game/utils";
-import { anchorPt } from "../../../gfx/anchor";
-import { drawTexture } from "../../../gfx/draw/drawTexture";
+import { drawSprite } from "../../../gfx/draw/drawSprite";
 import type { Texture } from "../../../gfx/gfx";
 import { Quad, quad, Rect, vec2 } from "../../../math/math";
 import { type Vec2 } from "../../../math/Vec2";
 import { _k } from "../../../shared";
 import type { Comp, GameObj, SpriteAnimPlayOpt } from "../../../types";
-import { warn } from "../../../utils/log";
 import { nextRenderAreaVersion } from "../physics/area";
 
 /**
@@ -349,101 +346,18 @@ export function sprite(
 
         draw(this: GameObj<SpriteComp>) {
             if (!spriteData) return;
-
-            const q = spriteData.frames[this.frame ?? 0];
-
-            if (!q) {
-                throw new Error(`Frame not found: ${this.frame ?? 0}`);
-            }
-
-            if (spriteData.slice9) {
-                // TODO: use scale or width / height, or both?
-                const { left, right, top, bottom, tileMode } =
-                    spriteData.slice9;
-
-                if (opt.tiled) {
-                    warn(
-                        "sprite(): 'tiled' option is ignored for 9-slice sprites. Use 'tileMode' in slice9 config instead.",
-                    );
-                }
-
-                const tw = spriteData.tex.width * q.w;
-                const th = spriteData.tex.height * q.h;
-                const iw = this.width - left - right;
-                const ih = this.height - top - bottom;
-                const w1 = left / tw;
-                const w3 = right / tw;
-                const w2 = 1 - w1 - w3;
-                const h1 = top / th;
-                const h3 = bottom / th;
-                const h2 = 1 - h1 - h3;
-                const quads = [
-                    // uv
-                    quad(0, 0, w1, h1),
-                    quad(w1, 0, w2, h1),
-                    quad(w1 + w2, 0, w3, h1),
-                    quad(0, h1, w1, h2),
-                    quad(w1, h1, w2, h2),
-                    quad(w1 + w2, h1, w3, h2),
-                    quad(0, h1 + h2, w1, h3),
-                    quad(w1, h1 + h2, w2, h3),
-                    quad(w1 + w2, h1 + h2, w3, h3),
-                    // transform
-                    quad(0, 0, left, top),
-                    quad(left, 0, iw, top),
-                    quad(left + iw, 0, right, top),
-                    quad(0, top, left, ih),
-                    quad(left, top, iw, ih),
-                    quad(left + iw, top, right, ih),
-                    quad(0, top + ih, left, bottom),
-                    quad(left, top + ih, iw, bottom),
-                    quad(left + iw, top + ih, right, bottom),
-                ];
-                const props = getRenderProps(this);
-                const offset = anchorPt(props.anchor || DEF_ANCHOR);
-                const offsetX = -(offset.x + 1) * 0.5 * this.width;
-                const offsetY = -(offset.y + 1) * 0.5 * this.height;
-                for (let i = 0; i < 9; i++) {
-                    const uv = quads[i];
-                    const transform = quads[i + 9];
-                    if (transform.w == 0 || transform.h == 0) {
-                        continue;
-                    }
-                    const isCenter = i === 4;
-                    const isEdge = !!(i & 1);
-                    const shouldTile = isCenter
-                        ? tileMode === "center" || tileMode === "all"
-                        : isEdge
-                        ? tileMode === "edges" || tileMode === "all"
-                        : false;
-                    drawTexture(
-                        Object.assign(props, {
-                            pos: transform.pos().add(offsetX, offsetY),
-                            anchor: "topleft",
-                            tex: spriteData.tex,
-                            quad: q.scale(uv),
-                            flipX: this.flipX,
-                            flipY: this.flipY,
-                            tiled: shouldTile,
-                            width: transform.w,
-                            height: transform.h,
-                        }),
-                    );
-                }
-            }
-            else {
-                drawTexture(
-                    Object.assign(getRenderProps(this), {
-                        tex: spriteData.tex,
-                        quad: q.scale(this.quad ?? new Quad(0, 0, 1, 1)),
-                        flipX: this.flipX,
-                        flipY: this.flipY,
-                        tiled: opt.tiled,
-                        width: this.width,
-                        height: this.height,
-                    }),
-                );
-            }
+            drawSprite(
+                Object.assign(getRenderProps(this), {
+                    frame: this.frame,
+                    sprite: spriteData,
+                    quad: this.quad,
+                    flipX: this.flipX,
+                    flipY: this.flipY,
+                    tiled: opt.tiled,
+                    width: this.width,
+                    height: this.height,
+                }),
+            );
         },
 
         add(this: GameObj<SpriteComp>) {
